+++ builds/backend/prefectFuncs.py
@@ -1,13 +1,11 @@
 from bson import ObjectId
 from uuid import UUID
 import traceback
-import httpx
 
 import db as flow_db
-from flows.AnalysisFlow import analysis_flows
+from AnalysisFlow import analysis_flows
 
 from prefect.deployments import run_deployment
-from prefect import serve, flow
 
 async def deploy_analysis(analysis_id: str, analysis_function: str):
     """
